+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Indicates additions to API, minor bug fixes and other small changes.
 
 0.1.1
 -----
+- bug:surface:win32: surface_set_name() would not clear previous name properly.
 - Makefile: fixed TARGET dependency on libcore
 - Makefile: added LIBCORE_SRC_PATH variable